feat: :sparkles: Added horizontal number picker and fixed unnecessary error popup
Now instead of inputs in settings to set the factor, road preset, etc. there's now a button that if clicked will popup a popup that allowes you to scroll between the options. Removed the device disconnected error from error code map because now I have the drop in message as error handler.