feat: :sparkles: Added personal device support, Added loading animation, Added heptic feedback
Added personal device support (when connecting to a device the first time it will become your personal device and when scanning for devices again, if it finds it it will connect to it, if the connect button is long pressed it will clear the personal device). Added loading animation that happens while the feedback is sent. Added short vibrations when a button is clicked or status of device changes in the home screen